dl/translation: move exception types to their own header
