+++ src/v/datalake/translation/BUILD
@@ -3,6 +3,7 @@ load("//bazel:build.bzl", "redpanda_cc_library")
 redpanda_cc_library(
     name = "model",
     hdrs = [
+        "errors.h",
         "types.h",
     ],
     include_prefix = "datalake/translation",
@@ -93,14 +94,15 @@ redpanda_cc_library(
         "record_multiplexer.h",
     ],
     implementation_deps = [
+        ":model",
         "//src/v/base",
         "//src/v/datalake:catalog_schema_manager",
-        "//src/v/iceberg/conversion:conversion_outcome",
         "//src/v/datalake:logger",
         "//src/v/datalake:record_schema_resolver",
         "//src/v/datalake:record_translator",
         "//src/v/datalake:table_creator",
         "//src/v/datalake:table_id_provider",
+        "//src/v/iceberg/conversion:conversion_outcome",
         "//src/v/storage:parser_utils",
         "//src/v/utils:lazy_abort_source",
         "//src/v/utils:prefix_logger",
@@ -156,6 +158,9 @@ redpanda_cc_library(
     ],
     hdrs = [
         "deps.h",
+    ],
+    implementation_deps = [
+        ":model",
     ],
     include_prefix = "datalake/translation",
     visibility = [":__subpackages__"],
