+++ src/v/datalake/translation/deps.cc
@@ -18,9 +18,10 @@
 #include "datalake/record_schema_resolver.h"
 #include "datalake/record_translator.h"
 #include "datalake/serde_parquet_writer.h"
+#include "datalake/translation/errors.h"
 #include "datalake/translation/state_machine.h"
-#include "datalake/translation/utils.h"
 #include "datalake/translation/translation_task.h"
+#include "datalake/translation/utils.h"
 #include "kafka/data/partition_proxy.h"
 #include "kafka/utils/txn_reader.h"
 #include "utils/human.h"
